Install sass and move css files into desktop styles folder
