+++ src/Components/Contact/Contact.js
@@ -1,4 +1,4 @@
-import './Contact.css';
+import '../../styles/desktop/contact.css';
 import valley from '../../assets/valley.jpeg';
 
 const Contact = () => {
